Added RearModule python file
Created a single file that will run all of the processes for the rear module. I combined the sensors and the camera code into one file and parallelized the processes so that they should run concurrently. I have never done parallel processing before so this may not work as I intended.